requirements.txt
torch>=2.0.0
torchvision>=0.15.0
facenet-pytorch>=2.5.3
wandb>=0.15.0
huggingface-hub>=0.16.0
torchmetrics>=1.0.0
torchmetrics[image]>=1.0.0
lpips>=0.1.4
torch-fidelity>=0.3.0
timm>=0.9.0
python-dotenv>=1.0.0
Pillow>=10.0.0 